Generation of random locations throughout the map + Scout reveals locations

+++ Unstable/StarFighter/StarFighter/Globals.cpp
@@ -116,7 +116,7 @@ float MinBetweenValues(sf::Vector2f values)
 
 size_t MinBetweenSizeTValues(size_t value1, size_t value2)
 {
-	float min = value1;
+	size_t min = value1;
 	if (value2 < min)
 	{
 		min = value2;
