+++ Unstable/StarFighter/StarFighter/Ore.cpp
@@ -28,7 +28,7 @@ Ore* Ore::CreateOre(string name)
 	new_ore->m_display_name = (*CurrentGame).m_oreConfig[name][OreData_Name];
 	new_ore->m_weight = (size_t)stoi((*CurrentGame).m_oreConfig[name][OreData_Weight]);
 	new_ore->m_price = (size_t)stoi((*CurrentGame).m_oreConfig[name][OreData_Price]);
-	new_ore->m_extraction_duration = atof((*CurrentGame).m_oreConfig[name][OreData_Extraction].c_str());
+	new_ore->m_extraction_duration = (float)atof((*CurrentGame).m_oreConfig[name][OreData_Extraction].c_str());
 	new_ore->m_propulsion = (size_t)stoi((*CurrentGame).m_oreConfig[name][OreData_Propulsion]);
 	new_ore->m_propulsion_speed_bonus = stof((*CurrentGame).m_oreConfig[name][OreData_PropulsionSpeedBonus]);
